+++ src/opportunities/pages/IndividualPost.tsx
